refactor: put cache in separate file
