+++ src/app/state/mod.rs
@@ -2,13 +2,11 @@
 
 pub mod action;
 mod datum;
+mod db;
 mod error;
-
-use std::{collections::HashMap, time::SystemTime};
 
 use dioxus::prelude::{use_context, use_context_provider, use_future, Scope};
 use dioxus_signals::Signal;
-use dioxus_std::storage::{new_storage, LocalStorage};
 use nix_health::NixHealth;
 use nix_rs::{
     flake::{url::FlakeUrl, Flake},
@@ -35,7 +33,7 @@ pub struct AppState {
     /// Cached [Flake] values indexed by [FlakeUrl]
     ///
     /// Most recently updated flakes appear first.
-    pub flake_cache: Signal<HashMap<FlakeUrl, Option<(SystemTime, Flake)>>>,
+    pub flake_cache: Signal<db::FlakeCache>,
 
     /// [Action] represents the next modification to perform on [AppState] signals
     pub action: Signal<(usize, Action)>,
@@ -45,13 +43,7 @@ impl AppState {
     fn new(cx: Scope) -> Self {
         tracing::info!("🔨 Creating new AppState");
         // TODO: Should we use new_synced_storage, instead? To allow multiple app windows?
-        let flake_cache = new_storage::<LocalStorage, _>(cx, "flake_cache".to_string(), || {
-            tracing::warn!("📦 No flake cache found");
-            FlakeUrl::suggestions()
-                .into_iter()
-                .map(|url| (url, None))
-                .collect()
-        });
+        let flake_cache = db::FlakeCache::new_signal(cx);
         AppState {
             flake_cache,
             ..AppState::default()
@@ -92,20 +84,6 @@ impl AppState {
         tracing::info!("setting flake url to {}", &url);
         self.flake_url.set(Some(url));
     }
-
-    /// Recently updated flakes, along with any unavailable flakes in cache.
-    pub fn recent_flakes(&self) -> Vec<FlakeUrl> {
-        let cache = &*self.flake_cache.read();
-        let mut pairs: Vec<_> = cache
-            .iter()
-            .filter_map(|(k, v)| v.as_ref().map(|(t, _)| (k, t)))
-            .collect();
-
-        // Sort by the timestamp in descending order.
-        pairs.sort_unstable_by(|a, b| b.1.cmp(a.1));
-
-        pairs.into_iter().map(|(k, _)| k.clone()).collect()
-    }
 }
 
 impl AppState {
@@ -130,11 +108,8 @@ impl AppState {
                     })
                     .await;
                     if let Some(Ok(flake)) = self.flake.read().current_value() {
-                        tracing::info!("Caching flake [{}]", flake_url);
                         self.flake_cache.with_mut(|cache| {
-                            let now = SystemTime::now();
-                            let v = Some((now, flake.clone()));
-                            cache.insert(flake_url, v);
+                            cache.update(flake_url, flake.clone());
                         });
                     }
                 }
@@ -146,10 +121,7 @@ impl AppState {
             // ... when URL changes.
             use_future(cx, (&flake_url,), |(flake_url,)| async move {
                 if let Some(flake_url) = flake_url {
-                    if let Some(Some((updated, cached_flake))) =
-                        self.flake_cache.read().get(&flake_url).cloned()
-                    {
-                        tracing::info!("Cache hit for {} (updated: {:?})", flake_url, updated);
+                    if let Some(cached_flake) = self.flake_cache.read().get(&flake_url) {
                         Datum::refresh_with(self.flake, async { Ok(cached_flake) }).await;
                     } else {
                         self.act(Action::RefreshFlake);
